feat(v2): update to lupdo v4
