+++ .github/workflows/coverage.yml
@@ -6,49 +6,45 @@ on:
       - main
 
 jobs:
-  build:
+  coverage:
     runs-on: ubuntu-latest
-
-    strategy:
-      fail-fast: false
-      matrix:
-        node-version: [18.x, 16.x]
-        database-type: [mysql8, 'mariadb1010']
-
     steps:
       - uses: actions/checkout@v3
 
-      - name: Use Node.js ${{ matrix.node-version }}
+      - name: Use Node.js
         uses: actions/setup-node@v3
         with:
-          node-version: ${{ matrix.node-version }}
+          node-version: 22.x
 
       - name: Start Database(s)
         run: |
-            docker-compose \
-              --file "docker-compose.yml" \
-              up \
-                --detach \
-                --build \
-                "${{ matrix.database-type }}"
+          docker-compose \
+            --file "docker-compose.yml" \
+            up \
+              --detach \
+              --build \
+              "mysql90"
 
       - name: Run npm install
         run: npm ci
 
       - name: Initialize Database
         run: |
-            docker-compose \
-              --file "docker-compose.yml" \
-              up \
-                "wait${{ matrix.database-type }}"
+          docker-compose \
+            --file "docker-compose.yml" \
+            up \
+              "waitmysql90"
 
-      - name: Build and test
-        run: |
-          npm run build
-          npm run test:coverage
+      - name: Build
+        run: npm run build
         env:
           CI: true
-          DB: ${{ matrix.database-type }}
+
+      - name: Coverage
+        run: npm run test:coverage
+        env:
+          CI: true
+          DB: mysql90
 
       - name: Code Coverage Report
         uses: irongut/CodeCoverageSummary@v1.3.0
@@ -64,7 +60,6 @@ jobs:
           thresholds: '75 90'
 
       - name: Upload coverage to Codecov
-        if:  ${{ matrix.database-type == 'mysql8' && matrix.node-version == '18.x'}}
         uses: codecov/codecov-action@v3
 
       - name: Stop Database(s)
