+++ .gitignore
@@ -103,6 +103,8 @@ dist
 # TernJS port file
 .tern-port
 
-
 # code-workspace
 *.code-workspace
+
+# api docs
+docs
