+++ .npmignore
@@ -1,5 +1,6 @@
 # Generated Files
 coverage/
+docs/
 
 # Source code
 src/
@@ -12,12 +13,13 @@ test/
 *.test.js
 
 # Configs
+.devcontainer/
 .github/
 .vscode/
 .travis.yml
 tsconfig.json
 typescript/
-.eslintrc.yml
+.eslintrc.js
 .prettierrc.yml
 .huskyrc.json
 .nvmrc
