--- .vscode/settings.json
@@ -1,24 +1,0 @@
-{
-    "eslint.validate": ["javascript", "typescript"],
-    "eslint.packageManager": "npm",
-    "[yaml]": {
-        "editor.formatOnSave": false
-    },
-    "[json]": {
-        "editor.formatOnSave": false
-    },
-    "[jsonc]": {
-        "editor.formatOnSave": false
-    },
-    "[javascript]": {
-        "editor.formatOnSave": false
-    },
-    "[typescript]": {
-        "editor.formatOnSave": false
-    },
-    "editor.codeActionsOnSave": {
-        "source.fixAll.eslint": true
-    },
-    "typescript.tsdk": "./node_modules/typescript/lib",
-    "jest.pathToJest": "npm run test --"
-}
