+++ src/mysql-connection.ts
@@ -1,14 +1,14 @@
 import { PdoConnection } from 'lupdo';
 import { Connection } from 'mysql2/promise';
 
-class MysqlConnection extends PdoConnection {
-    constructor(public readonly connection: Connection) {
-        super();
-    }
+export class MysqlConnection extends PdoConnection {
+  constructor(public readonly connection: Connection) {
+    super();
+  }
 
-    async query(sql: string): Promise<void> {
-        await this.connection.query(sql);
-    }
+  async query(sql: string): Promise<void> {
+    await this.connection.query(sql);
+  }
 }
 
 export default MysqlConnection;
