+++ typescript/tsconfig.cjs.json
@@ -1,18 +1,10 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
-    "moduleResolution": "node",
     "declaration": false,
     "sourceMap": false,
     "incremental": false,
-    "target": "ES2020",
-    "module": "ES2020",
-    "outDir": "../dist/esm",
-    "lib": [
-      "ES2020",
-      "DOM"
-    ],
-    "downlevelIteration": true
+    "outDir": "../dist/cjs"
   },
   "include": [
     "../src/"
